requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-client-cache==0.11.1
aiosignal==1.4.0
aiosqlite==0.20.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
cachetools==6.2.2
catalogue==2.0.10
cattrs==25.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudscraper==1.2.71
confection==0.1.5
courlan==1.3.2
cssselect==1.3.0
curl_cffi==0.14.0
cymem==2.0.13
dateparser==1.2.2
deepdiff==8.6.1
defusedxml==0.7.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
fastapi==0.128.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
htmldate==1.9.4
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
inscriptis==2.7.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
murmurhash==1.0.15
narwhals==2.15.0
networkx==3.6
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.3.5
openbb==4.6.0
openbb-benzinga==1.5.1
openbb-bls==1.2.1
openbb-cftc==1.2.1
openbb-commodity==1.4.2
openbb-congress-gov==1.1.1
openbb-core==1.5.8
openbb-crypto==1.5.1
openbb-currency==1.5.1
openbb-derivatives==1.5.1
openbb-econdb==1.4.1
openbb-economy==1.5.1
openbb-equity==1.5.1
openbb-etf==1.5.1
openbb-federal-reserve==1.5.1
openbb-fixedincome==1.5.1
openbb-fmp==1.5.2
openbb-fred==1.5.1
openbb-government-us==1.5.2
openbb-imf==2.0.0
openbb-index==1.5.1
openbb-intrinio==1.5.1
openbb-news==1.5.1
openbb-oecd==1.5.1
openbb-platform-api==1.2.3
openbb-polygon==1.5.1
openbb-regulators==1.5.1
openbb-sec==1.5.1
openbb-tiingo==1.5.1
openbb-tradingeconomics==1.5.1
openbb-us-eia==1.2.1
openbb-yfinance==1.5.2
openpyxl==3.1.5
orderly-set==5.5.0
packaging==25.0
pandas==2.3.3
peewee==3.18.3
pgvector==0.4.1
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
preshed==3.0.12
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
random-user-agent==1.0.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-cache==1.2.1
requests-file==3.0.1
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
ruff==0.13.3
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.50.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.57.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
url-normalize==1.4.3
urllib3==2.5.0
uuid7==0.1.0
uvicorn==0.40.0
wasabi==1.1.3
weasel==0.4.3
webencodings==0.5.1
websockets==15.0.1
wrapt==2.0.1
xlrd==2.0.2
xmltodict==0.13.0
xx_ent_wiki_sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.8.0/xx_ent_wiki_sm-3.8.0-py3-none-any.whl#sha256=6f3c4b853852ea9e9d2dc76cc950dddb10a7e4c42d813308caefe6c5e8be2f0a
yarl==1.22.0
yfinance==0.2.66
zipp==3.23.0
